Added functionality to push the updated configuration of components to it's linked components
