+++ viewer-config-persistence/src/test/java/nl/b3p/viewer/config/app/ApplicationTest.java
@@ -124,20 +124,4 @@ public class ApplicationTest extends TestUtil {
             assert (false);
         }
     }
-
-    @Test
-    public void testUpdateComponentsInMotherApplication() throws Exception{
-        initData(true);
-        Application mashup = app.createMashup("mashup",  entityManager,true);
-
-        String differentString = "different";
-        for (ConfiguredComponent component : app.getComponents()) {
-            component.setConfig(component.getConfig() + differentString);
-        }
-
-        for (ConfiguredComponent component : mashup.getComponents()) {
-            assertTrue(component.getConfig().contains(differentString));
-        }
-    }
-
 }
